🚿 clean up utils
