+++ modules/utils/src/env.ts
@@ -1,0 +1,5 @@
+export const isNode = () =>
+  typeof process !== "undefined" &&
+  typeof process.versions !== "undefined" &&
+  typeof process.versions.node !== "undefined";
+
