+++ modules/utils/src/index.ts
@@ -1,10 +1,13 @@
-export * from "./client";
+export * from "./addresses";
+export * from "./comparisons";
 export * from "./crypto";
 export * from "./currency";
+export * from "./delay";
+export * from "./env";
+export * from "./hexStrings";
 export * from "./identifiers";
+export * from "./json";
 export * from "./logger";
 export * from "./math";
-export * from "./misc";
-export * from "./node";
 export * from "./store";
-export * from "./validation";
+export * from "./validator";
